Create EmailReceiver.java
Added the email alias functionality to the EmailReceiver class. Added a list of email aliases and a method to add or remove them. Right now i'm just adding them in the notification GUI class but you'll be able to call the method in anything else to add or remove an email.